remove bench_test
