+++ test/manager_test.exs
@@ -1,8 +1,6 @@
 defmodule ManagerTest do
   use ExUnit.Case, async: false
 
-  alias Features.Manager.TestSpider
-
   setup do
     Application.put_env(:crawly, :concurrent_requests_per_domain, 1)
     Application.put_env(:crawly, :closespider_itemcount, 10)
@@ -20,7 +18,7 @@ defmodule ManagerTest do
 
     on_exit(fn ->
       :meck.unload()
-      Crawly.Engine.stop_spider(TestSpider)
+      Crawly.Engine.stop_spider(Manager.TestSpider)
       Application.put_env(:crawly, :manager_operations_timeout, 30_000)
       Application.put_env(:crawly, :concurrent_requests_per_domain, 1)
       Application.put_env(:crawly, :closespider_timeout, 20)
@@ -29,16 +27,16 @@ defmodule ManagerTest do
   end
 
   test "max request per minute is respected" do
-    :ok = Crawly.Engine.start_spider(TestSpider)
+    :ok = Crawly.Engine.start_spider(Manager.TestSpider)
 
-    {:stored_requests, num} = Crawly.RequestsStorage.stats(TestSpider)
+    {:stored_requests, num} = Crawly.RequestsStorage.stats(Manager.TestSpider)
     assert num == 1
     Process.sleep(1_00)
 
-    {:stored_items, num} = Crawly.DataStorage.stats(TestSpider)
+    {:stored_items, num} = Crawly.DataStorage.stats(Manager.TestSpider)
     assert num == 1
 
-    :ok = Crawly.Engine.stop_spider(TestSpider)
+    :ok = Crawly.Engine.stop_spider(Manager.TestSpider)
     assert %{} == Crawly.Engine.running_spiders()
   end
 
@@ -47,7 +45,7 @@ defmodule ManagerTest do
 
     Application.put_env(:crawly, :manager_operations_timeout, 50)
     Application.put_env(:crawly, :closespider_itemcount, 1)
-    :ok = Crawly.Engine.start_spider(TestSpider)
+    :ok = Crawly.Engine.start_spider(Manager.TestSpider)
 
     assert_receive :itemcount_timeout
 
@@ -55,6 +53,7 @@ defmodule ManagerTest do
   end
 
   test "Closespider timeout is respected" do
+
     Process.register(self(), :spider_closed_callback_test)
 
     # Ignore closespider_itemcount
@@ -63,24 +62,73 @@ defmodule ManagerTest do
     Application.put_env(:crawly, :closespider_timeout, 10)
 
     Application.put_env(:crawly, :manager_operations_timeout, 50)
-    :ok = Crawly.Engine.start_spider(TestSpider)
+    :ok = Crawly.Engine.start_spider(Manager.TestSpider)
 
     assert_receive :itemcount_timeout
     assert %{} == Crawly.Engine.running_spiders()
   end
 
   test "Can't start already started spider" do
-    :ok = Crawly.Engine.start_spider(TestSpider)
+    :ok = Crawly.Engine.start_spider(Manager.TestSpider)
 
     assert {:error, :spider_already_started} ==
-             Crawly.Engine.start_spider(TestSpider)
+             Crawly.Engine.start_spider(Manager.TestSpider)
   end
 
   test "Spider closed callback is called when spider is stopped" do
     Process.register(self(), :spider_closed_callback_test)
-    :ok = Crawly.Engine.start_spider(TestSpider)
-    :ok = Crawly.Engine.stop_spider(TestSpider, :manual_stop)
+    :ok = Crawly.Engine.start_spider(Manager.TestSpider)
+    :ok = Crawly.Engine.stop_spider(Manager.TestSpider, :manual_stop)
 
     assert_receive :manual_stop
+  end
+end
+
+defmodule Manager.TestSpider do
+  use Crawly.Spider
+
+  def override_settings() do
+    on_spider_closed_callback = fn reason ->
+      case Process.whereis(:spider_closed_callback_test) do
+        nil ->
+          :nothing_to_do
+
+        _pid ->
+          send(:spider_closed_callback_test, reason)
+      end
+    end
+
+    [on_spider_closed_callback: on_spider_closed_callback]
+  end
+
+  def base_url() do
+    "https://www.example.com"
+  end
+
+  def init() do
+    [
+      start_urls: ["https://www.example.com/blog.html"]
+    ]
+  end
+
+  def parse_item(_response) do
+    path = Enum.random(1..100)
+
+    %{
+      :items => [
+        %{title: "t_#{path}", url: "example.com", author: "Me", time: "not set"}
+      ],
+      :requests => [
+        Crawly.Utils.request_from_url("https://www.example.com/#{path}")
+      ]
+    }
+  end
+
+  def spider_closed(:manual_stop) do
+    send(:spider_closed_callback_test, :manual_stop)
+  end
+
+  def spider_closed(_) do
+    :ignored
   end
 end
